+++ common/src/main/java/dev/ithundxr/createnumismatics/config/CClient.java
@@ -27,6 +27,7 @@ public class CClient extends ConfigBase {
     public final ConfigGroup client = group(0, "client", Comments.client);
 
     public final ConfigBool scalePonderGui = b(true, "scalePonderGUI", Comments.scalePonderGui);
+    public final ConfigEnum<EmojiMode> coinEmojiPrefix = e(EmojiMode.DEFAULT, "coinEmojiPrefix", Comments.coinEmojiPrefix);
 
     // Based off of https://github.com/Layers-of-Railways/Railway/blob/68713f0fbb20080b7e207c070b1595bdbbc1bc00/common/src/main/java/com/railwayteam/railways/config/CClient.java
 
@@ -38,5 +39,19 @@ public class CClient extends ConfigBase {
     private static class Comments {
         static final String client = "Client-only settings - If you're looking for general settings, look inside your worlds serverconfig folder!";
         static final String scalePonderGui = "Whether GUIs in Ponders should be scaled down";
+        static final String coinEmojiPrefix = "Depending on the mode, a bevel icon can be created using :bevel:, :coin-bevel:, or :numi-bevel: respectively.";
+    }
+
+    public enum EmojiMode {
+        DEFAULT(""),
+        PREFIX_COIN("coin-"),
+        PREFIX_NUMI("numi-")
+        ;
+
+        public final String prefix;
+
+        EmojiMode(String prefix) {
+            this.prefix = prefix;
+        }
     }
 }